+++ src/transform/plugins/term/termDefinitions.ts
